Code formatting and time complexity details
